+++ test/options.spec.js
@@ -1,61 +1,59 @@
 'use strict';
 
-var express = require('express');
-var chai = require('chai');
-var expect = chai.expect;
-var GridFsStorage = require('../index');
-var setting = require('./utils/settings');
-var uploads = require('./utils/uploads');
-var request = require('supertest');
-var multer = require('multer');
-var md5File = require('md5-file');
-var path = require('path');
-var crypto = require('crypto');
-var mute = require('mute');
-var sinon = require('sinon');
-var sinonChai = require('sinon-chai');
+const express = require('express');
+const chai = require('chai');
+const expect = chai.expect;
+const GridFsStorage = require('../index');
+const setting = require('./utils/settings');
+const uploads = require('./utils/uploads');
+const request = require('supertest');
+const multer = require('multer');
+const md5File = require('md5-file');
+const path = require('path');
+const crypto = require('crypto');
+const mute = require('mute');
+const sinon = require('sinon');
 
 chai.use(require('chai-interface'));
-
-chai.use(sinonChai);
+chai.use(require('sinon-chai'));
 
 describe('module usage', function () {
   this.timeout(4000);
-  var result, app,
+  let result, app, storage,
     db, gfs, spy, logSpy;
   
-  before(function () {
+  before(() => {
     spy = sinon.spy();
     logSpy = sinon.spy();
     app = express();
   });
   
   describe('all configuration options', function () {
-    var messages = [];
-    before(function (done) {
-      var counter = 0;
-      var storage = GridFsStorage({
+    const messages = [];
+    before((done) => {
+      let counter = 0;
+      storage = GridFsStorage({
         url: setting.mongoUrl(),
-        filename: function (req, file, cb) {
-          crypto.randomBytes(16, function (err, raw) {
+        filename: (req, file, cb) => {
+          crypto.randomBytes(16, (err, raw) => {
             cb(err, err ? undefined : raw.toString('hex') + path.extname(file.originalname));
           });
         },
-        metadata: function (req, file, cb) {
+        metadata: (req, file, cb) => {
           cb(null, req.body);
         },
-        identifier: function (req, file, cb) {
+        identifier: (req, file, cb) => {
           cb(null, Math.floor(Math.random() * 1000000));
         },
-        chunkSize: function (req, file, cb) {
+        chunkSize: (req, file, cb) => {
           cb(null, Math.floor(Math.random() * 20000) + 10000);
         },
-        root: function (req, file, cb) {
-          var roots = ['myfiles', 'otherfiles'];
+        root: (req, file, cb) => {
+          const roots = ['myfiles', 'otherfiles'];
           counter++;
           cb(null, counter === 1 ? roots[0] : roots[1]);
         },
-        log: function (err, log) {
+        log: (err, log) => {
           logSpy(err, log);
           if (err) {
             return messages.push({ type: 'error', data: err });
@@ -67,15 +65,13 @@ describe('module usage', function () {
       
       storage.on('file', spy);
       
-      var upload = multer({
-        storage: storage
-      });
+      const upload = multer({ storage });
       
-      app.post('/opts', upload.array('photos', 2), function (req, res) {
+      app.post('/opts', upload.array('photos', 2), (req, res) => {
         res.send({ headers: req.headers, files: req.files, body: req.body });
       });
       
-      storage.once('connection', function (grid, database) {
+      storage.once('connection', (grid, database) => {
         gfs = grid;
         db = database;
         
@@ -84,7 +80,7 @@ describe('module usage', function () {
           .attach('photos', uploads.files[0])
           .attach('photos', uploads.files[1])
           .field('field', 'field')
-          .end(function (err, res) {
+          .end((err, res) => {
             result = res.body;
             done();
           });
@@ -94,21 +90,21 @@ describe('module usage', function () {
     });
     
     it('should use a 16 bytes hexadecimal name with an extension', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file).to.have.property('filename').that.matches(/^[a-f0-9]{32}\.jpg$/);
       });
       done();
     });
     
     it('should have a metadata property with the value {field: "field"}', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file).to.have.property('metadata').that.is.eql({ field: 'field' });
       });
       done();
     });
     
     it('should have a id property that matches an ObjectId format', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file).to.have.property('id').that.is.a('string');
         expect(file).to.have.property('id').that.matches(/^[a-f0-9]{24}$/);
       });
@@ -116,7 +112,7 @@ describe('module usage', function () {
     });
     
     it('should have a id property that matches an ObjectId format', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file).to.have.property('id').that.is.a('string');
         expect(file).to.have.property('id').that.matches(/^[a-f0-9]{24}$/);
       });
@@ -124,7 +120,7 @@ describe('module usage', function () {
     });
     
     it('should have a grid property with the stored file info', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file).to.have.property('grid').that.have.interface({
           chunkSize: Number,
           contentType: String,
@@ -139,7 +135,7 @@ describe('module usage', function () {
     });
     
     it('should have the same MD5 signature than the upload', function (done) {
-      result.files.forEach(function (file, index) {
+      result.files.forEach((file, index) => {
         expect(file.grid.md5).to.be.equal(md5File(uploads.files[index]));
       });
       done();
@@ -155,63 +151,42 @@ describe('module usage', function () {
     });
     
     it('should have a different chunkSize between 10000 and 30000', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file.grid.chunkSize).to.be.within(10000, 30000);
       });
       done();
     });
     
     it('should be stored under a different root', function (done) {
-      db.collections().then(function (collections) {
+      db.collections().then((collections) => {
         expect(collections).to.have.length(5);
-        collections.forEach(function (col) {
+        collections.forEach((col) => {
           expect(['system.indexes', 'myfiles.files', 'myfiles.chunks', 'otherfiles.files', 'otherfiles.chunks']).to.include(col.collectionName);
         });
         done();
       });
     });
-    
-    after(function (done) {
-      db.collection('myfiles.chunks').drop()
-        .then(function () {
-          return db.collection('myfiles.files').drop();
-        })
-        .then(function () {
-          return db.collection('otherfiles.chunks').drop();
-        })
-        .then(function () {
-          return db.collection('otherfiles.files').drop();
-        })
-        .then(function () {
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-    });
   });
   
   describe('fixed value configuration options', function () {
-    var unmute;
+    let unmute;
     
-    before(function (done) {
+    before((done) => {
       unmute = mute();
-      var storage = GridFsStorage({
+      storage = GridFsStorage({
         url: setting.mongoUrl(),
         chunkSize: 131072,
         root: 'myfiles',
         log: true
       });
       
-      var upload = multer({
-        storage: storage
-      });
+      const upload = multer({ storage });
       
-      app.post('/fixed', upload.array('photos', 2), function (req, res) {
+      app.post('/fixed', upload.array('photos', 2), (req, res) => {
         res.send({ headers: req.headers, files: req.files, body: req.body });
       });
       
-      storage.once('connection', function (grid, database) {
+      storage.once('connection', (grid, database) => {
         gfs = grid;
         db = database;
         
@@ -220,7 +195,7 @@ describe('module usage', function () {
           .attach('photos', uploads.files[0])
           .attach('photos', uploads.files[1])
           .field('field', 'field')
-          .end(function (err, res) {
+          .end((err, res) => {
             result = res.body;
             unmute();
             done();
@@ -229,20 +204,20 @@ describe('module usage', function () {
     });
     
     it('should have a different fixed chunkSize with the value 131072', function (done) {
-      result.files.forEach(function (file) {
+      result.files.forEach((file) => {
         expect(file.grid.chunkSize).to.be.equal(131072);
       });
       done();
     });
     
     it('should be stored under a different root with the value myfiles', function (done) {
-      db.collection('myfiles.files', { strict: true }, function (err) {
+      db.collection('myfiles.files', { strict: true }, (err) => {
         expect(err).to.be.equal(null);
-        db.collection('myfiles.chunks', { strict: true }, function (err) {
+        db.collection('myfiles.chunks', { strict: true }, (err) => {
           expect(err).to.be.equal(null);
-          db.collection('fs.files', { strict: true }, function (err) {
+          db.collection('fs.files', { strict: true }, (err) => {
             expect(err).not.to.be.equal(null);
-            db.collection('fs.chunks', { strict: true }, function (err) {
+            db.collection('fs.chunks', { strict: true }, (err) => {
               expect(err).not.to.be.equal(null);
               done();
             });
@@ -250,44 +225,29 @@ describe('module usage', function () {
         });
       });
     });
-    
-    after(function (done) {
-      db.collection('myfiles.chunks').drop()
-        .then(function () {
-          return db.collection('myfiles.files').drop();
-        })
-        .then(function () {
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-    });
   });
   
   describe('failed request', function () {
-    var unmute;
-    before(function (done) {
+    let unmute;
+    before((done) => {
       unmute = mute();
-      var storage = GridFsStorage({
+      storage = GridFsStorage({
         url: setting.mongoUrl()
       });
       
-      var upload = multer({
-        storage: storage
-      });
+      const upload = multer({ storage });
       
-      app.post('/fail', upload.array('photos', 1), function (req, res) {
+      app.post('/fail', upload.array('photos', 1), (req, res) => {
         res.send({ headers: req.headers, file: req.file, body: req.body });
       });
       
-      storage.once('connection', function () {
+      storage.once('connection', () => {
         request(app)
           .post('/fail')
           .attach('photos', uploads.files[0])
           .attach('photos', uploads.files[1])
           .field('field', 'field')
-          .end(function (err, res) {
+          .end((err, res) => {
             result = res.body;
             unmute();
             done();
@@ -296,19 +256,22 @@ describe('module usage', function () {
     });
     
     it('should fail with an error', function (done) {
-      gfs.files.count({}, function (err, count) {
+      gfs.files.count({}, (err, count) => {
         expect(count).to.equal(0);
         done(err);
       });
     });
   });
   
-  after(function () {
-    return db
-      .dropDatabase()
-      .then(function () {
-        return db.close(true);
-      });
+  afterEach(() => {
+    if (storage) {
+      storage.removeAllListeners();
+    }
+    if (db) {
+      return db
+        .dropDatabase()
+        .then(() => db.close(true));
+    }
   });
   
 });
